edit Time Format,Page New
